// src/app/features/students/model/student.model.ts
export interface Student {
  id: string;
  name: string;
  email: string;
  cpf: string;
  imageUrl: string;
  age: number;
  timePratice: number;
  graduation: string;
  paymentDetail: number;
  details: string;
  lastGraduationDetail: string;
  montlyPayment: boolean;
  isActivate: boolean;
  last_graduation_date: Date;
  payment_date: Date;
  altered_at: Date;
  created_at: Date;
}